start over with inference
This is a better design but still not sufficient; as I worked through it
I realized that the type inference step isn't compatible with typeless
lambdas, because you can't infer-forward what the types of the arguments
are unless the thing has a type notated on it. Without that, I can't
think of a way to avoid full HM, or at least some backtracking.