instantiate all base types on first use
list:Nil wasn't being correctly typed because its type, list:List
list:a, was never being instantiated because it never appears in the
head of an application. Instead, we were generating substitutions for
list:a and causing the whole module to explode. This correctly
instantiates that, at the cost of many extra instantiations for each
module.